+++ terra_geocrud/views.py
@@ -125,6 +125,9 @@ class CrudFeatureViewSet(ReversionMixin, FeatureViewSet):
         feature_name = feature.layer.crud_view.get_feature_title(feature)
         date_formatted = formats.date_format(timezone.localtime(), "SHORT_DATETIME_FORMAT")
         new_name = f"{template.name}_{feature_name}_{date_formatted}{suffix}"
+        # TODO: fix 'using' key at None. Let django auto detect goog template engine
+        #  (for example, by creating a specific folder for odt and docx templates
+        #  OR passign OdtEngine and DocxEngine BEFORE DjangoTemplateEngine
         response = TemplateResponse(
             request=self.request,
             template=template.template_file.name,